requirements.txt
setuptools
-e .
fastapi
uvicorn
langchain
langchain-openai
langchain_community
python-dotenv
faiss-cpu
pypdf